User Reg Page
its 1am and im hungry. But redid the register page so it looks pretty and added the necessary fields. I also broke the verification for all the fields except *DOB Required* but i'm on 6 hours sleep and i cannot be bothered figuring out how to ffix it. Happy days. xoxo Love Eric